# docs/swagger.yaml
...
swagger: "2.0"
info:
  description: "Тестовое задание для реализации проекта \"Форумы\" на курсе по баз\
    ам данных в\nТехнопарке Mail.ru (https://park.mail.ru).\n"
  version: "0.1.0"
  title: "forum"
host: "virtserver.swaggerhub.com"
basePath: "/paulnopaul/tbdb/0.1.0"
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /forum/create:
    post:
      summary: "Создание форума"
      description: "Создание нового форума.\n"
      operationId: "forumCreate"
      parameters:
      - in: "body"
        name: "forum"
        description: "Данные форума."
        required: true
        schema:
          $ref: "#/definitions/Forum"
        x-exportParamName: "Forum"
      responses:
        "201":
          description: "Форум успешно создан.\nВозвращает данные созданного форума\
            .\n"
          schema:
            $ref: "#/definitions/Forum"
        "404":
          description: "Владелец форума не найден.\n"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Форум уже присутсвует в базе данных.\nВозвращает данные ра\
            нее созданного форума.\n"
          schema:
            $ref: "#/definitions/Forum"
  /forum/{slug}/details:
    get:
      summary: "Получение информации о форуме"
      description: "Получение информации о форуме по его идентификаторе.\n"
      operationId: "forumGetOne"
      consumes: []
      parameters:
      - name: "slug"
        in: "path"
        description: "Идентификатор форума."
        required: true
        type: "string"
        format: "identity"
        x-exportParamName: "Slug"
      responses:
        "200":
          description: "Информация о форуме.\n"
          schema:
            $ref: "#/definitions/Forum"
        "404":
          description: "Форум отсутсвует в системе.\n"
          schema:
            $ref: "#/definitions/Error"
  /forum/{slug}/create:
    post:
      summary: "Создание ветки"
      description: "Добавление новой ветки обсуждения на форум.\n"
      operationId: "threadCreate"
      parameters:
      - name: "slug"
        in: "path"
        description: "Идентификатор форума."
        required: true
        type: "string"
        format: "identity"
        x-exportParamName: "Slug"
      - in: "body"
        name: "thread"
        description: "Данные ветки обсуждения."
        required: true
        schema:
          $ref: "#/definitions/Thread"
        x-exportParamName: "Thread"
      responses:
        "201":
          description: "Ветка обсуждения успешно создана.\nВозвращает данные созда\
            нной ветки обсуждения.\n"
          schema:
            $ref: "#/definitions/Thread"
        "404":
          description: "Автор ветки или форум не найдены.\n"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Ветка обсуждения уже присутсвует в базе данных.\nВозвращае\
            т данные ранее созданной ветки обсуждения.\n"
          schema:
            $ref: "#/definitions/Thread"
  /forum/{slug}/users:
    get:
      summary: "Пользователи данного форума"
      description: "Получение списка пользователей, у которых есть пост или ветка\
        \ обсуждения в данном форуме.\n\nПользователи выводятся отсортированные по\
        \ nickname в порядке возрастания.\nПорядок сотрировки должен соответсвоват\
        ь побайтовому сравнение в нижнем регистре.\n"
      operationId: "forumGetUsers"
      consumes: []
      parameters:
      - name: "slug"
        in: "path"
        description: "Идентификатор форума."
        required: true
        type: "string"
        format: "identity"
        x-exportParamName: "Slug"
      - name: "limit"
        in: "query"
        description: "Максимальное кол-во возвращаемых записей."
        required: false
        type: "number"
        default: 100.0
        maximum: 10000
        minimum: 1
        format: "int32"
        x-exportParamName: "Limit"
        x-optionalDataType: "Float32"
      - name: "since"
        in: "query"
        description: "Идентификатор пользователя, с которого будут выводиться поль\
          зоватли\n(пользователь с данным идентификатором в результат не попадает\
          ).\n"
        required: false
        type: "string"
        format: "identity"
        x-exportParamName: "Since"
        x-optionalDataType: "String"
      - name: "desc"
        in: "query"
        description: "Флаг сортировки по убыванию.\n"
        required: false
        type: "boolean"
        x-exportParamName: "Desc"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "Информация о пользователях форума.\n"
          schema:
            $ref: "#/definitions/Users"
        "404":
          description: "Форум отсутсвует в системе.\n"
          schema:
            $ref: "#/definitions/Error"
  /forum/{slug}/threads:
    get:
      summary: "Список ветвей обсужления форума"
      description: "Получение списка ветвей обсужления данного форума.\n\nВетви об\
        суждения выводятся отсортированные по дате создания.\n"
      operationId: "forumGetThreads"
      consumes: []
      parameters:
      - name: "slug"
        in: "path"
        description: "Идентификатор форума."
        required: true
        type: "string"
        format: "identity"
        x-exportParamName: "Slug"
      - name: "limit"
        in: "query"
        description: "Максимальное кол-во возвращаемых записей."
        required: false
        type: "number"
        default: 100.0
        maximum: 10000
        minimum: 1
        format: "int32"
        x-exportParamName: "Limit"
        x-optionalDataType: "Float32"
      - name: "since"
        in: "query"
        description: "Дата создания ветви обсуждения, с которой будут выводиться з\
          аписи\n(ветвь обсуждения с указанной датой попадает в результат выборки\
          ).\n"
        required: false
        type: "string"
        format: "date-time"
        x-exportParamName: "Since"
        x-optionalDataType: "Time"
      - name: "desc"
        in: "query"
        description: "Флаг сортировки по убыванию.\n"
        required: false
        type: "boolean"
        x-exportParamName: "Desc"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "Информация о ветках обсуждения на форуме.\n"
          schema:
            $ref: "#/definitions/Threads"
        "404":
          description: "Форум отсутсвует в системе.\n"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Can't find user with id #42\n"
        description: "Текстовое описание ошибки.\nВ процессе проверки API никаких\
          \ проверок на содерижимое данного описание не делается.\n"
        readOnly: true
...
  User:
    type: "object"
    required:
    - "email"
    - "fullname"
    properties:
      nickname:
        type: "string"
        format: "identity"
        example: "j.sparrow"
        description: "Имя пользователя (уникальное поле).\nДанное поле допускает т\
          олько латиницу, цифры и знак подчеркивания.\nСравнение имени регистронез\
          ависимо.\n"
        readOnly: true
      fullname:
        type: "string"
        example: "Captain Jack Sparrow"
        description: "Полное имя пользователя."
        x-isnullable: false
      about:
        type: "string"
        format: "text"
        example: "This is the day you will always remember as the day that you almost\
          \ caught Captain Jack Sparrow!"
        description: "Описание пользователя."
      email:
        type: "string"
        format: "email"
        example: "captaina@blackpearl.sea"
        description: "Почтовый адрес пользователя (уникальное поле)."
        x-isnullable: false
    description: "Информация о пользователе.\n"
    example:
      nickname: "j.sparrow"
      about: "This is the day you will always remember as the day that you almost\
        \ caught Captain Jack Sparrow!"
      fullname: "Captain Jack Sparrow"
      email: "captaina@blackpearl.sea"
  Users:
    type: "array"
    items:
      $ref: "#/definitions/User"
...
  Forum:
    type: "object"
    required:
    - "slug"
    - "title"
    - "user"
    properties:
      title:
        type: "string"
        example: "Pirate stories"
        description: "Название форума."
        x-isnullable: false
      user:
        type: "string"
        format: "identity"
        example: "j.sparrow"
        description: "Nickname пользователя, который отвечает за форум."
        x-isnullable: false
      slug:
        type: "string"
        format: "identity"
        example: "pirate-stories"
        description: "Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL),\
          \ уникальное поле."
        pattern: "^(\\d|\\w|-|_)*(\\w|-|_)(\\d|\\w|-|_)*$"
        x-isnullable: false
      posts:
        type: "number"
        format: "int64"
        example: 200000.0
        description: "Общее кол-во сообщений в данном форуме.\n"
        readOnly: true
      threads:
        type: "number"
        format: "int32"
        example: 200.0
        description: "Общее кол-во ветвей обсуждения в данном форуме.\n"
        readOnly: true
    description: "Информация о форуме.\n"
    example:
      threads: 200.0
      title: "Pirate stories"
      user: "j.sparrow"
      posts: 200000.0
      slug: "pirate-stories"
  Thread:
    type: "object"
    required:
    - "author"
    - "message"
    - "title"
    properties:
      id:
        type: "number"
        format: "int32"
        example: 42.0
        description: "Идентификатор ветки обсуждения."
        readOnly: true
      title:
        type: "string"
        example: "Davy Jones cache"
        description: "Заголовок ветки обсуждения."
        x-isnullable: false
      author:
        type: "string"
        format: "identity"
        example: "j.sparrow"
        description: "Пользователь, создавший данную тему."
        x-isnullable: false
      forum:
        type: "string"
        format: "identity"
        example: "pirate-stories"
        description: "Форум, в котором расположена данная ветка обсуждения."
        readOnly: true
      message:
        type: "string"
        format: "text"
        example: "An urgent need to reveal the hiding place of Davy Jones. Who is\
          \ willing to help in this matter?"
        description: "Описание ветки обсуждения."
        x-isnullable: false
      votes:
        type: "number"
        format: "int32"
        description: "Кол-во голосов непосредственно за данное сообщение форума."
        readOnly: true
      slug:
        type: "string"
        format: "identity"
        example: "jones-cache"
        description: "Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL).\n\
          В данной структуре slug опционален и не может быть числом.\n"
        readOnly: true
        pattern: "^(\\d|\\w|-|_)*(\\w|-|_)(\\d|\\w|-|_)*$"
      created:
        type: "string"
        format: "date-time"
        example: "2017-01-01T00:00:00.000Z"
        description: "Дата создания ветки на форуме."
        x-isnullable: true
    description: "Ветка обсуждения на форуме.\n"
    example:
      forum: "pirate-stories"
      author: "j.sparrow"
      created: "2017-01-01T00:00:00.000Z"
      votes: 0.80082819046101150206595775671303272247314453125
      id: 42.0
      title: "Davy Jones cache"
      message: "An urgent need to reveal the hiding place of Davy Jones. Who is willing\
        \ to help in this matter?"
      slug: "jones-cache"
  Threads:
    type: "array"
    items:
      $ref: "#/definitions/Thread"
...